pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dtiam"
version = "3.0.0"
description = "Dynatrace IAM CLI - A kubectl-style CLI for Dynatrace Identity and Access Management"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Timothy Stewart" }
]
keywords = ["dynatrace", "iam", "cli", "identity", "access", "management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "platformdirs>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.8",
    "ruff>=0.2",
    "types-PyYAML>=6.0",
]

[project.scripts]
dtiam = "dtiam.cli:main"

[project.urls]
Homepage = "https://github.com/jtimothystewart/Python-IAM-Utility-3.0"
Repository = "https://github.com/jtimothystewart/Python-IAM-Utility-3.0"

[tool.hatch.build.targets.wheel]
packages = ["src/dtiam"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["dtiam"]

[tool.mypy]
python_version = "3.10"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
